+++ Core/Constants.lua
@@ -2,7 +2,7 @@
 local ADDON_NAME, CVarMaster = ...
 
 CVarMaster.Constants = {
-    VERSION = "1.1.0",
+    VERSION = "1.2.0",
 
     -- Display modes
     MODES = {